# [REVIEW] RF: fix stream bug causing performance regressions